use uint16 for player game stats (#2293)

+++ app/models/colyseus-models/player.ts
@@ -84,9 +84,9 @@ export default class Player extends Schema implements IPlayer {
   @type("string") map: DungeonPMDO
   @type({ set: "string" }) effects: Effects = new Effects()
   @type(["string"]) regionalPokemons = new ArraySchema<Pkm>()
-  @type("uint8") rerollCount: number = 0
-  @type("uint8") totalMoneyEarned: number = 0
-  @type("uint8") totalPlayerDamageDealt: number = 0
+  @type("uint16") rerollCount: number = 0
+  @type("uint16") totalMoneyEarned: number = 0
+  @type("uint16") totalPlayerDamageDealt: number = 0
   @type("float32") eggChance: number = 0
   commonRegionalPool: Pkm[] = new Array<Pkm>()
   uncommonRegionalPool: Pkm[] = new Array<Pkm>()
